% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{AccuProb}
\alias{AccuProb}
\title{Convert node abundance to edge abundance}
\usage{
AccuProb(P, Tree, r = FALSE)
}
\arguments{
\item{P}{A numeric matrix or data frame. Compositional data. Row represents nodes on the phylogenetic tree. Columns represents samples.}

\item{Tree}{\code{phylo} class. Phylogenetic tree.}

\item{r}{If calculation is being performed on variance or values.}
}
\value{
A numeric matrix or data frame. Compositional data. Row represents edges on the phylogenetic tree. Columns represent samples.
}
\description{
Convert node abundance to edge abundance
}
