% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAFOTPlus.R
\name{ConIndDAFOT}
\alias{ConIndDAFOT}
\title{Phylogenetic Conditional Independence Test with Rank Correlation (ConIndDAFOT)}
\usage{
ConIndDAFOT(
  P,
  Y,
  X,
  Tree,
  condgen = NULL,
  ExY = NULL,
  ExX = NULL,
  Exk = min(10, length(ExY)),
  method = "Dn",
  neighbor = min(10, length(Y)),
  step = 200
)
}
\arguments{
\item{P}{A numeric matrix or data frame. Compositional data. Row represents nodes on the phylogenetic tree. Column represents samples.}

\item{Y}{A vector. The interested variable.}

\item{X}{The confounding variables. Row represents samples. Each column corresponds to each confounding variable.}

\item{Tree}{\code{phylo} class. Phylogenetic tree.}

\item{condgen}{\code{function}. Conditional distribution generating function.}

\item{ExY}{Extra data generated from the conditional distribution of Y given X.}

\item{ExX}{Extra data generated from the distribution of X.}

\item{Exk}{the number of neighbors to draw Y from the neighbors of X.}

\item{method}{Dn or Rn or Tn. Dn is the Hoeffding's D test; Rn is the Blum-Kiefer-Rosenblatt's R; Tn is the Bergsma-Dassios-Yanaginoto's tau test.}

\item{neighbor}{The number of neighbors to estimate local rank correlation.}

\item{step}{Permutation times.}
}
\value{
a \code{list} with components:
\describe{
\item{Stat}{a vector with length 2. The first result is the correlation coefficient using the weighted sum approach; the second result is the correlation coefficient using maximum approach.}
\item{P}{a vector with length 2. The first result is the obtained p-value using the weighted sum approache; the second result is the obtained p-value using maximum approach.}
}
}
\description{
A phylogenetic independence test directly measures the association between the outcome of interest and the total microbial abundance with adjustment of confounding effects.
}
\details{
This method is designed to control confounding effect in the basis of IndDAFOT. Similar to the idea of the nearest neighbor mehtod, we consider the nearest neighbor conditional rank correlation coefficient for each lineage. And conditional rank correlations for lineage are aggregated as the weighted sum and maximum to capture signals.
}
\examples{
library(ape)
Tree <- rtree(100)
P <- rbind(
  matrix(1, nrow = length(Tree$tip.label), ncol = 100),
  matrix(0, nrow = Tree$Nnode, ncol = 100)
)
Y <- c(rep(1, 50), rep(2, 50))
X <- matrix(rnorm(500), nrow = 100)
n <- 1
ConIndDAFOT(P, Y, X, Tree, step = n)
}
