% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{EdgeLExtract}
\alias{EdgeLExtract}
\title{Extract edge weights from the phylogenetic tree}
\usage{
EdgeLExtract(Tree)
}
\arguments{
\item{Tree}{\code{phylo} class. Phylogenetic tree.}
}
\value{
A vector of numerics representing edge weights on the tree.
}
\description{
Extract edge weights from the phylogenetic tree
}
