% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAFOTPlus_function.R
\name{IndDAFOT}
\alias{IndDAFOT}
\title{Phylogenetic Independence Test with Rank Correlation (IndDAFOT)}
\usage{
IndDAFOT(P, Y, tree, method = "Dn", step = 200)
}
\arguments{
\item{P}{A numeric matrix or data frame. Compositional data. Row represents nodes on the phylogenetic tree. Column represents samples.}

\item{Y}{A vector. The interested variable.}

\item{tree}{\code{phylo} class. Phylogenetic tree.}

\item{method}{Dn or Rn or Tn. Dn is the Hoeffding's D test; Rn is the Blum-Kiefer-Rosenblatt's R; Tn is the Bergsma-Dassios-Yanaginoto's tau test.}

\item{step}{Permutation times.}
}
\value{
a \code{list} with components:
\itemize{
\item{ \code{Stat}, a vector with length 2. The first result is the correlation coefficient using the weighted sum approach; the second result is the correlation coefficient using maximum approach.}
\item{ \code{P}, a vector with length 2. The first result is the obtained p-value using the weighted sum approache; the second result is the obtained p-value using maximum approach.}
}
}
\description{
A phylogenetic independence test directly measures the association between the outcome of interest and the total microbial abundance in each lineage (subtree) by a rank correlation.
}
\details{
This method is designed to capture possible nonlinear associations between the outcome of interest and microbial abundance. It aggeregates the rank correlations for different lineages in two ways: the weighted sum and maximum.
}
\examples{
library(ape)
Tree <- rtree(100)
P <- rbind(matrix(1, nrow = length(Tree$tip.label), ncol = 100), matrix(0, nrow = Tree$Nnode, ncol = 100))
Y <- c(rep(1, 50), rep(2, 50))
IndDAFOT(P, Y, Tree)
}
