% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{NodeDepth}
\alias{NodeDepth}
\title{Return the depth of each node on the tree}
\usage{
NodeDepth(Tree)
}
\arguments{
\item{Tree}{\code{phylo} class. Phylogenetic tree.}
}
\value{
An integer vector containing the depth of each node.
}
\description{
Return the depth of each node on the tree
}
