% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{standEdge}
\alias{standEdge}
\title{Standardize edge weights}
\usage{
standEdge(EdgeP, EdgeQ, nP, nQ)
}
\arguments{
\item{EdgeP}{A numeric matrix or data frame. Compositional data. Row represents edges on the phylogenetic tree. Columns represent samples.}

\item{EdgeQ}{A numeric matrix or data frame. Compositional data. Row represents edges on the phylogenetic tree. Columns represent samples.}

\item{nP}{The number of columns in EdgeP}

\item{nQ}{The number of columns in EdgeQ}
}
\value{
A numeric vector. The differences between input edge matrixes.
}
\description{
Standardize edge weights
}
