% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_ConIndDAFOT}
\alias{test_ConIndDAFOT}
\title{A data set designed for testing the ConIndDAFOT function}
\format{
## `test_ConIndDAFOT`
A named list:
\itemize{
  \item {P} {Relative abundances}
  \item {Y} {Outcome}
  \item {X} {Confounding effects}
  \item {tree} {Tree}
  \item {condgen} {}
  \item {ExY} {}
  \item {ExX} {}
  \item {Exk} {}
}
}
\usage{
test_ConIndDAFOT
}
\description{
A data set designed for testing the ConIndDAFOT function
}
\keyword{datasets}
