% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_DAFOT}
\alias{test_DAFOT}
\title{A data set designed for testing the DAFOT function}
\format{
## `test_DAFOT`
A named list:
\itemize{
  \item {P} {First group}
  \item {Q} {Second group}
  \item {tree} {Tree}
}
}
\usage{
test_DAFOT
}
\description{
A data set designed for testing the DAFOT function
}
\keyword{datasets}
