% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_IndDAFOT}
\alias{test_IndDAFOT}
\title{A data set designed for testing the IndDAFOT function}
\format{
## `test_IndDAFOT`
A named list:
\itemize{
  \item{P}{Relative abundances}
  \item{Y}{Outcome}
  \item{tree}{Tree}
}
}
\usage{
test_IndDAFOT
}
\description{
A data set designed for testing the IndDAFOT function
}
\keyword{datasets}
