% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{tree.corr.cond}
\alias{tree.corr.cond}
\title{Conditional correlation between one single edge and Y}
\usage{
tree.corr.cond(EdgeP, Y, X, method, neighbor = min(10, length(Y)), weight = F)
}
\arguments{
\item{EdgeP}{A numeric matrix or data frame. Compositional data. Row represents edges on the phylogenetic tree. Columns represent samples.}

\item{Y}{A vector. The interested variable.}

\item{X}{The confounding variables. Row represents samples. Each column corresponds to each confounding variable.}

\item{method}{Dn or Rn or Tn. Dn is the Hoeffding's D test; Rn is the Blum-Kiefer-Rosenblatt's R; Tn is the Bergsma-Dassios-Yanaginoto's tau test.}

\item{neighbor}{The number of neighbors to estimate local rank correlation.}

\item{weight}{A vector. Weights to apply to the test results.}
}
\value{
A vector. First field is the weighted sum of test results. Second is the max result.
}
\description{
Conditional correlation between one single edge and Y
}
