% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{tree.corr}
\alias{tree.corr}
\title{Correlation between one single edge and Y}
\usage{
tree.corr(EdgeP, Y, method, weight = F)
}
\arguments{
\item{EdgeP}{A numeric matrix or data frame. Compositional data. Row represents edges on the phylogenetic tree. Columns represent samples.}

\item{Y}{A vector. The interested variable.}

\item{method}{Dn or Rn or Tn. Dn is the Hoeffding's D test; Rn is the Blum-Kiefer-Rosenblatt's R; Tn is the Bergsma-Dassios-Yanaginoto's tau test.}

\item{weight}{A vector. Weights to apply to the test results.}
}
\value{
A vector. First field is the weighted sum of test results. Second is the max result.
}
\description{
Correlation between one single edge and Y
}
