% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAFOT_function.R
\name{DAFOT}
\alias{DAFOT}
\title{Detector of Active Flow on a Tree}
\usage{
DAFOT(P, Q, Tree, times, alpha)
}
\arguments{
\item{P}{A matrix. Compositional data of 1st group, each row corresponds to each node of tree and each column corresponds to each sample.}

\item{Q}{A matrix. Compositional data of 2nd group, each row corresponds to each node of tree and each column corresponds to each sample.}

\item{Tree}{A phylo class. Phylogenetic tree.}

\item{times}{A integer. The times of permutation test.}

\item{alpha}{A numeric. Significance level}
}
\value{
\code{DAFOT} returns an object of class \code{dafot}.

An object of class \code{dafot} is a list containing following components:
\itemize{
\item \code{Stat} the maximum of test statistics
\item \code{PValue} P value calculated from permutation test
\item \code{StatThre} the threshold for alpha level test
\item \code{ActiveEdge} the set of active edges
}
}
\description{
\code{DAFOT} is a funtion used to detect the difference between two group of compositional data.
}
\details{
This function is developed based on an optimal transport perspective. It can help detect the specific location of difference at a tree.
}
\examples{
library(ape)
Tree <- rtree(100)
alphaP <- c(rep(1, length(Tree$tip.label)), rep(0, Tree$Nnode))
alphaQ <- c(rep(1, length(Tree$tip.label)), rep(0, Tree$Nnode))
DataPQ <- DataGenerating(100, 100, alphaP, alphaQ, 1000)
DAFOT(DataPQ$P, DataPQ$Q, Tree, 100, 0.05)

}
\seealso{
\code{\link{SCalculation}}
}
\author{
Shulei Wang
}
