% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{EdgeLExtract}
\alias{EdgeLExtract}
\title{Extract sorted edge weights from the phylogenetic tree
Add zero for the root node edge length
Return all 0's if tree is invalid}
\usage{
EdgeLExtract(Tree)
}
\arguments{
\item{Tree}{\code{phylo} class. Phylogenetic tree.}
}
\value{
A vector of numerics representing edge weights on the tree.
}
\description{
Extract sorted edge weights from the phylogenetic tree
Add zero for the root node edge length
Return all 0's if tree is invalid
}
